Option to add list of foldername mapping
Sometimes SharePoint sync gives different foldernames to the main folder. Example sometimes the Documents folder is named "XDrive Documents", but on another PC it may be named "XDrive Documenten"
The app allows to do mapping of foldernames. This is only applicable to the initial folder, not the subfolders. It's a list of mappings.

Example line: "XDrive Documents=XDrive Documenten"

When doing a backup:
 - When a folder in the source is named "XDrive Documents", then the file folderlist.json will be named "XDrive Documenten" and also in the backup there will be a folder with name "XDrive Documenten"
 - So when we reuse a hard disk with backups taken on another PC, and the folder on that disk is named "XDrive Documenten", while on the new PC, "